Renamed file from mypage_contianer to MyPageContainer
